Move implementation to client too
